requirements.txt
keyboard==0.13.5
Pillow==10.0.1
pygame==2.5.2
pystray==0.19.5
six==1.16.0
